Fix console logs and refactor date to export function instead of composable
