+++ src/helpers/date.ts
@@ -1,0 +1,10 @@
+export const formatDateForBackend = (date = new Date()): string => {
+  const yyyy = date.getFullYear()
+  const mm = String(date.getMonth() + 1).padStart(2, '0') // months are 0-based
+  const dd = String(date.getDate()).padStart(2, '0')
+  const hh = String(date.getHours()).padStart(2, '0')
+  const mi = String(date.getMinutes()).padStart(2, '0')
+  const ss = String(date.getSeconds()).padStart(2, '0')
+
+  return `${yyyy}-${mm}-${dd} ${hh}:${mi}:${ss}`
+}
