+++ src/views/OrderView.vue
@@ -12,7 +12,7 @@ import { useOrderItems } from '@/API/OrderItems.ts'
 import ConfirmDialog from 'primevue/confirmdialog'
 import { useConfirm } from 'primevue/useconfirm'
 import type { OrderItem } from '@/types/OrderItem.ts'
-import { useDateFormatter } from '@/Composables/DateFormatter.ts'
+import { useDateFormatter } from '@/helpers/date.ts'
 
 const { getOrderItems, deleteOrderItem, createOrderItem, updateOrderItem } = useOrderItems()
 const orderItems = ref([])
